Adding automated anvil testing client creation.
