--- packages/protocol-sdk/.env.anvil
@@ -1,2 +1,0 @@
-FORK_RPC_URL="https://rpc.zora.co/"
-FORK_BLOCK_NUMBER=6133407
