+++ packages/protocol-sdk/src/premint/premint-client.ts
@@ -20,8 +20,8 @@ import type {
 } from "./premint-api-client";
 import { PremintAPIClient } from "./premint-api-client";
 import type { DecodeEventLogReturnType } from "viem";
-import { ClientBase, REWARD_PER_TOKEN } from "./client-base";
-import { OPEN_EDITION_MINT_SIZE } from "./constants";
+import { ClientBase, REWARD_PER_TOKEN } from "../apis/client-base";
+import { OPEN_EDITION_MINT_SIZE } from "../constants";
 
 type MintArgumentsSettings = {
   tokenURI: string;
