+++ packages/protocol-sdk/tsconfig.json
@@ -21,5 +21,5 @@
     "outDir": "dist"
   },
   "exclude": ["node_modules/**", "dist/**"],
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "vitest.config.ts"]
 }
